postgresql/bookstore-recommendations-demo/requirements.txt
streamlit==1.28.1
psycopg2-binary==2.9.7
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
plotly==5.17.0 
watchdog==6.0.0